chain added and tested
