+++ lib/app.js
@@ -1,7 +1,1 @@
-let files = require("./blockChain.js");
-console.log(files);
-
-block = new files.Block("KIM");
-console.log(block);
-chain = new files.Chain();
-console.log(chain);
+let blockChain = require("./blockChain.js");
