+++ lib/blockChain.js
@@ -1,24 +1,4 @@
-const ENCRYPT = require('crypto-js/sha512');
-const DATEFORMAT = require("dateformat");
-
-class Block {
-  constructor(vote, previousHash = "") {
-    this.index = 1;
-    this.timestamp = this.currentDate();
-    this.vote = vote;
-    this.previousHash = previousHash;
-    this.hash = this.calculateHash();
-  }
-
-  calculateHash() {
-    return ENCRYPT(this.index + this.previousHash + this.timestamp + JSON.stringify(this.vote)).toString();
-  }
-
-  currentDate() {
-    return DATEFORMAT(new Date(), "isoDateTime");
-  }
-  
-}
+const block = require("./block.js");
 
 class Blockchain {
   constructor() {
@@ -26,7 +6,7 @@ class Blockchain {
   }
 
   createFirstBlock() {
-    return new Block('Trumpyboy');
+    return new block.Block('Trumpyboy');
   }
 
   findLastBlock() {
@@ -44,4 +24,4 @@ class Blockchain {
 }
 
 exports.Chain = Blockchain;
-exports.Block = Block;
+
