+++ spec/chain_spec.js
@@ -1,0 +1,5 @@
+let Chain = require('../lib/blockChain.js');
+
+describe('block creation', function () {
+
+});
